+++ src/3_branch.js
@@ -1126,6 +1126,8 @@ if (CORDOVA_BUILD) { // jshint undef:false
  * The `forgetHide` property defaults to false, and when set to true, will forget if the user has opened the banner previously, and thus will always show the banner to them even if they have closed it in the past. It can also be set to an integer, in which case, it would forget that the user has previously hid the banner after X days.
  * The `position` property, defaults to 'top', but can also be set to 'bottom' if you would prefer to show the app banner from the bottom of the screen.
  * The `customCSS` property allows you to style the banner, even if it is isolated within an iframe. To assist you with device specific styles, the body element of the banner has one of three classes: `branch-banner-android`, `branch-banner-ios`, or `branch-banner-desktop`.
+ * The `mobileSticky` property defaults to false, but can be set to true if you want the user to continue to see the app banner as they scroll.
+ * The `desktopSticky` property defaults to true, but can be set to false if you want the user to only see the app banner when they are scrolled to the top of the page.
  * ```js
  * branch.banner(
  *     {
@@ -1138,6 +1140,8 @@ if (CORDOVA_BUILD) { // jshint undef:false
  *          disableHide: true,
  *          forgetHide: true, // Can also be set to an integer. For example: 10, would forget that the user previously hid the banner after 10 days
  *          position: 'bottom',
+ *          mobileSticky: true,
+ *          desktopSticky: true,
  *          customCSS: '.title { color: #F00; }'
  *     },
  *     {... link data ...}
@@ -1217,6 +1221,8 @@ if (WEB_BUILD) { // jshint undef:false
  *     disableHide: false,                     // Should the user have the ability to hide the banner? (show's X on left side)
  *     forgetHide: false,                      // Should we show the banner after the user closes it? Can be set to true, or an integer to show again after X days
  *     position: 'top',                        // Sets the position of the banner, options are: 'top' or 'bottom', and the default is 'top'
+ *     mobileSticky: false,                    // Determines whether the mobile banner will be set `position: fixed;` (sticky) or `position: absolute;`, defaults to false *this property only applies when the banner position is 'top'
+ *     desktopSticky: true,                    // Determines whether the desktop banner will be set `position: fixed;` (sticky) or `position: absolute;`, defaults to true *this property only applies when the banner position is 'top'
  *     customCSS: '.title { color: #F00; }',   // Add your own custom styles to the banner that load last, and are gauranteed to take precedence, even if you leave the banner in an iframe
  *     make_new_link: false                    // Should the banner create a new link, even if a link already exists?
  * }, {
@@ -1271,6 +1277,8 @@ if (WEB_BUILD) { // jshint undef:false
 			forgetHide: typeof options['forgetHide'] == 'number' ? options['forgetHide'] : !!options['forgetHide'],
 			position: options['position'] || 'top',
 			customCSS: options['customCSS'] || '',
+			mobileSticky: typeof options['mobileSticky'] == 'undefined' ? false : options['mobileSticky'],
+			desktopSticky: typeof options['desktopSticky'] == 'undefined' ? true : options['desktopSticky'],
 			make_new_link: !!options['make_new_link']
 		};
 		if (typeof options['showMobile'] != 'undefined') {